backend/services/llm-python-rpc/requirements.txt
grpcio>=1.75.1
protobuf>=5.29.3
